Trait chiquito::compiler::FixedGenContext

pub trait FixedGenContext<F> {
    fn assign(&mut self, offset: usize, lhs: Queriable<F>, rhs: F);
}

A trait that represents a fixed column generation context. It provides an interface for assigning values to fixed columns in a circuit at the specified offset.

Required Methods

fn assign(&mut self, offset: usize, lhs: Queriable<F>, rhs: F)

Takes a Queriable object representing the fixed column (lhs) and the value (rhs) to be assigned.