Module chiquito::dsl

Re-exports

pub use sc::*;

Modules

cb

lb

sc

Structs

CircuitContext

A generic structure designed to handle the context of a circuit for generic types F, TraceArgs and StepArgs. The struct contains a Circuit instance and implements methods to build the circuit, add various components, and manipulate the circuit. F is a generic type representing the field of the circuit. TraceArgs is a generic type representing the arguments passed to the trace function. StepArgs is a generic type representing the arguments passed to the step_type_def function.

StepTypeContext

A generic structure designed to handle the context of a step type definition. The struct contains a StepType instance and implements methods to build the step type, add components, and manipulate the step type. F is a generic type representing the field of the step type. Args is the type of the step instance witness generation arguments.

StepTypeHandler

StepTypeSetupContext

StepTypeWGHandler

Enums

StepTypeDefInput

Functions

circuit

Creates a Circuit instance by providing a name and a definition closure that is applied to a mutable CircuitContext. The user customizes the definition closure by calling CircuitContext functions. This is the main function that users call to define a Chiquito circuit. Currently, the name is not used for annotation within the function, but it may be used in future implementations.