Struct chiquito::frontend::dsl::CircuitContext

pub struct CircuitContext<F, TraceArgs> { /* private fields */ }

A generic structure designed to handle the context of a circuit for generic types F, TraceArgs and StepArgs. The struct contains a Circuit instance and implements methods to build the circuit, add various components, and manipulate the circuit. F is a generic type representing the field of the circuit. TraceArgs is a generic type representing the arguments passed to the trace function. StepArgs is a generic type representing the arguments passed to the step_type_def function.

Implementations

impl<F, TraceArgs> CircuitContext<F, TraceArgs>

pub fn forward(&mut self, name: &str) -> Queriable<F>

Adds a forward signal to the circuit with a name string and zero rotation and returns a Queriable instance representing the added forward signal.

pub fn forward_with_phase(&mut self, name: &str, phase: usize) -> Queriable<F>

Adds a forward signal to the circuit with a name string and a specified phase and returns a Queriable instance representing the added forward signal.

pub fn shared(&mut self, name: &str) -> Queriable<F>

Adds a shared signal to the circuit with a name string and zero rotation and returns a Queriable instance representing the added shared signal.

pub fn shared_with_phase(&mut self, name: &str, phase: usize) -> Queriable<F>

Adds a shared signal to the circuit with a name string and a specified phase and returns a Queriable instance representing the added shared signal.

pub fn fixed(&mut self, name: &str) -> Queriable<F>

pub fn expose(&mut self, queriable: Queriable<F>, offset: ExposeOffset)

Exposes the first step instance value of a forward signal as public.

pub fn import_halo2_advice( &mut self, name: &str, column: Halo2Column<Advice> ) -> Queriable<F>

Imports a halo2 advice column with a name string into the circuit and returns a Queriable instance representing the imported column.

pub fn import_halo2_fixed( &mut self, name: &str, column: Halo2Column<Fixed> ) -> Queriable<F>

Imports a halo2 fixed column with a name string into the circuit and returns a Queriable instance representing the imported column.

pub fn step_type(&mut self, name: &str) -> StepTypeHandler

Adds a new step type with the specified name to the circuit and returns a StepTypeHandler instance. The StepTypeHandler instance can be used to define the step type using the step_type_def function.

pub fn step_type_def<D, Args, S: Into<StepTypeDefInput>, R>( &mut self, step: S, def: D ) -> StepTypeWGHandler<F, Args, R>where D: FnOnce(&mut StepTypeContext<F>) -> StepTypeWGHandler<F, Args, R>, R: Fn(&mut StepInstance<F>, Args) + 'static,

Defines a step type using the provided StepTypeHandler and a function that takes a mutable reference to a StepTypeContext. This function typically adds constraints to a step type and defines witness generation.

pub fn trace<D>(&mut self, def: D)where D: Fn(&mut TraceContext<F>, TraceArgs) + 'static,

Sets the trace function that builds the witness. The trace function is responsible for adding step instances defined in step_type_def. The function is entirely left for the user to implement and is Turing complete. Users typically use external parameters of type TraceArgs to generate cell values for witness generation, and call the add function to add step instances with witness values.

pub fn new_table(&self, table: LookupTableStore<F>) -> LookupTable

pub fn pragma_first_step<STH: Into<StepTypeHandler>>(&mut self, step_type: STH)

Enforce the type of the first step by adding a constraint to the circuit. Takes a StepTypeHandler parameter that represents the step type.

pub fn pragma_last_step<STH: Into<StepTypeHandler>>(&mut self, step_type: STH)

Enforce the type of the last step by adding a constraint to the circuit. Takes a StepTypeHandler parameter that represents the step type.

pub fn pragma_num_steps(&mut self, num_steps: usize)

pub fn pragma_disable_q_enable(&mut self)

impl<F: Field + Hash, TraceArgs> CircuitContext<F, TraceArgs>

pub fn fixed_gen<D>(&mut self, def: D)where D: Fn(&mut FixedGenContext<F>) + 'static,

Executes the fixed generation function provided by the user and sets the fixed assignments for the circuit. The fixed generation function is responsible for assigning fixed values to fixed columns. It is entirely left for the user to implement and is Turing complete. Users typically generate cell values and call the assign function to fill the fixed columns.